requirements.txt
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.2
timm==0.9.2
Pillow==10.0.1
tqdm==4.66.1
pyyaml==6.0.1
openpyxl==3.1.5